+++ dirección/src/main/java/com/example/direccion/repository/RegionRepository.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Repository;
 
 import com.example.direccion.model.Region;
 @Repository
-public interface RegionRepository extends JpaRepository<Region, Long> {
-
+public interface RegionRepository extends JpaRepository<Region, Integer> {
+    
+    
 }
